refactor(storage): extract MediaStoreDownloader and satisfy lint gates
Move the HTTP download streaming out of MediaStoreFileOperationsImpl to
keep the class within size limits, chain the original exception into the
download failure error, apply ktlint formatting, and fix an illegal
character in a test method name.
